Fairly extreme refactor
I don't know why - but the script before was annoying me when debugging
and it just seemed to have lots of stuff. So I went a bit overboard
and pulled everything libraryish.

I think the key refactor was
stars -> star_count
star -> event

Probably a bit too much work but it is neat now